+++ app/src/main/java/com/trever/android/data/auth/AuthInterceptor.kt
@@ -1,54 +1,93 @@
-package com.trever.android.data.auth // 실제 패키지 경로로 수정해주세요
+package com.trever.android.data.auth
 
-import android.util.Log // Log 임포트
-import kotlinx.coroutines.flow.first
+
+import com.trever.android.data.remote.AuthApi
+import com.trever.android.data.remote.RefreshRequest
 import kotlinx.coroutines.runBlocking
+import okhttp3.Authenticator
 import okhttp3.Interceptor
+import okhttp3.Request
 import okhttp3.Response
-//import javax.inject.Inject // 만약 Hilt/Dagger를 사용한다면, 아니면 생성자 주입으로 TokenStore 받기
+import okhttp3.Route
 
-// Hilt/Dagger 사용하지 않을 경우, 생성자로 TokenStore를 직접 받습니다.
-// class AuthInterceptor @Inject constructor(private val tokenStore: TokenStore) : Interceptor {
-class AuthInterceptor(private val tokenStore: TokenStore) : Interceptor {
-
+/**
+ * 매 요청에 Authorization 헤더를 붙여주는 Interceptor
+ */
+class AuthInterceptor(
+    private val tokenStore: TokenStore
+) : Interceptor {
     override fun intercept(chain: Interceptor.Chain): Response {
-        // <<< 로그 추가: 인터셉터 실행 시작 >>>
-        Log.d("AuthInterceptor", "AuthInterceptor 실행됨.")
+        val original = chain.request()
+        val access = runBlocking { tokenStore.getAccessToken() }
 
-        val accessToken =
-            runBlocking {
-                // <<< 로그 추가: 토큰 가져오기 시도 및 결과 >>>
-                Log.d("AuthInterceptor", "TokenStore에서 AccessToken 가져오기 시도...")
-                // 만약 tokenStore.getAccessToken()이 Flow<String?>을 반환하고 .first()를 사용해야 한다면,
-                // 다음 라인의 주석을 풀고 아래 'val token =' 라인은 주석 처리합니다.
-                // tokenStore.getAccessToken().first().also { tokenValue -> Log.d("AuthInterceptor", "가져온 AccessToken (from flow): $tokenValue") }
+        val req: Request = if (!access.isNullOrBlank()) {
+            original.newBuilder()
+                .header("Authorization", "Bearer $access")
+                .build()
+        } else original
 
-                // 만약 tokenStore.getAccessToken()이 직접 String?을 반환한다면 (예: SharedPreferences 직접 접근)
-                val token = tokenStore.getAccessToken() // 여기가 Flow<String?> 이라면 .first()가 필요할 수 있습니다.
-                Log.d("AuthInterceptor", "가져온 AccessToken: $token")
-                token // 반환
+        return chain.proceed(req)
+    }
+}
+
+/**
+ * 401 발생 시 자동으로 refresh 토큰을 사용해 토큰을 재발급하고
+ * 원래 요청을 재시도하는 Authenticator
+ */
+class TokenAuthenticator(
+    private val tokenStore: TokenStore,
+    private val authApi: AuthApi,
+    private val onTokenExpired: () -> Unit
+) : Authenticator {
+
+    override fun authenticate(route: Route?, response: Response): Request? {
+        if (responseCount(response) >= 2) return null
+
+        val currentRefreshToken = runBlocking { tokenStore.getRefreshToken() }
+        if (currentRefreshToken.isNullOrBlank()) {
+            runBlocking { tokenStore.clear() }
+            onTokenExpired() // 로그인 화면 이동 등 처리
+            return null
+        }
+
+        synchronized(this) {
+            val newAccessToken = runBlocking { tokenStore.getAccessToken() }
+            val originalRequestAccessToken = response.request.header("Authorization")?.substringAfter("Bearer ")
+            if (newAccessToken != null && originalRequestAccessToken != newAccessToken) {
+                return response.request.newBuilder()
+                    .header("Authorization", "Bearer $newAccessToken")
+                    .build()
             }
-        val originalRequest = chain.request()
-        val requestBuilder = originalRequest.newBuilder()
 
-        if (!accessToken.isNullOrBlank()) {
-            // <<< 로그 추가: 토큰 발견 및 헤더 추가 시도 >>>
-            Log.d("AuthInterceptor", "AccessToken 발견됨 ('$accessToken'). 헤더에 추가합니다: Bearer $accessToken")
-            requestBuilder.header("Authorization", "Bearer $accessToken")
-        } else {
-            // <<< 로그 추가: 토큰 없음 >>>
-            Log.d("AuthInterceptor", "AccessToken이 null이거나 비어있습니다. 헤더에 추가하지 않습니다.")
+            val newTokens = runCatching {
+                runBlocking {
+                    authApi.refresh(RefreshRequest(currentRefreshToken))
+                }
+            }.getOrNull()?.data
+
+            if (newTokens == null) {
+                runBlocking { tokenStore.clear() }
+                onTokenExpired() // 로그인 화면 이동 등 처리
+                return null
+            }
+
+            runBlocking {
+                tokenStore.saveTokens(newTokens.accessToken, newTokens.refreshToken)
+            }
+
+            return response.request.newBuilder()
+                .header("Authorization", "Bearer ${newTokens.accessToken}")
+                .build()
         }
+    }
 
-        val request = requestBuilder.build()
-        // <<< 로그 추가: 최종 요청 정보 및 헤더 목록 >>>
-        Log.d("AuthInterceptor", "최종 요청 URL: ${request.url}")
-        Log.d("AuthInterceptor", "--- 최종 요청 헤더 목록 시작 ---")
-        request.headers.forEach { header ->
-            Log.d("AuthInterceptor", "헤더: ${header.first} = ${header.second}")
+    private fun responseCount(response: Response): Int {
+        var res = response
+        var count = 1
+        while (res.priorResponse != null) {
+            count++
+            res = res.priorResponse!!
         }
-        Log.d("AuthInterceptor", "--- 최종 요청 헤더 목록 끝 ---")
-
-        return chain.proceed(request)
+        return count
     }
 }
