+++ app/src/main/java/com/trever/android/data/remote/VehicleMappers.kt
@@ -2,18 +2,13 @@ package com.trever.android.data.remote
 
 import com.trever.android.domain.model.AuctionCar
 import com.trever.android.domain.model.SearchCarItem
-import com.trever.android.domain.model.SellerInfo
 import com.trever.android.domain.model.Tag
 import com.trever.android.domain.model.VehicleDetail
 import com.trever.android.domain.model.VehicleSummary
 import com.trever.android.ui.auction.AuctionDetailUi
-import com.trever.android.ui.auction.BidUi
 import com.trever.android.ui.auction.SellerUi
+import com.trever.android.ui.utils.formatKoreanWon
 import com.trever.android.ui.utils.formatMileage
-
-import kotlin.div
-
-import kotlin.toString
 import java.text.NumberFormat
 import java.util.Locale
 
@@ -42,27 +37,7 @@ fun VehicleDetail.toBuyDetailUi(): AuctionDetailUi {
     )
 }
 
-fun VehicleDto.toDomain(): VehicleSummary = VehicleSummary(
-    id = id,
-    carName = carName ?: "",
-    manufacturer = manufacturer ?: "",
-    model = model ?: "",
-    year = year_value ?: 0,
-    mileageKm = mileage ?: 0,
-    transmission = transmission ?: "",
-    fuelType = fuelType ?: "",
-    priceWon = price ?: 0L,
-    isAuction = isAuction?.equals("Y", ignoreCase = true) ?: false, // null 체크 추가
-    auctionId = auctionId ?: 0,
-    imageUrl = representativePhotoUrl ?: "",
-    locationAddress = locationAddress ?: "",
-    favoriteCount = favoriteCount ?: 0,
-    createdAt = createdAt ?: "",
-    vehicleTypeName = vehicleTypeName ?: "",
-    mainOptions = mainOptions ?: emptyList(),
-    totalOptionsCount = totalOptionsCount ?: 0,
-    liked = isFavorite ?: false // isFavorite 필드 매핑 추가
-)
+
 
 fun VehicleDto.toAuctionCar(): AuctionCar {
     return AuctionCar(
@@ -122,13 +97,7 @@ data class VehiclesPageDto(
     val pageNumber: Int,
     val pageSize: Int
 )
-fun VehiclesPageDto.toDomain(): Page<VehicleSummary> =
-    Page(
-        items = vehicles.map { it.toDomain() },
-        totalCount = totalCount,
-        pageNumber = pageNumber,
-        pageSize = pageSize
-    )
+
 
 fun VehicleDetailResponse.toVehicleDetail(): VehicleDetail {
     return VehicleDetail(
@@ -231,90 +200,7 @@ fun Vehicle.toSearchCarItem(): SearchCarItem =
         )
     }
 
-//fun VehicleDetail.toBuyDetailUi(): AuctionDetailUi {
-//    val priceValue = price
-//    val priceText = formatKoreanWon(priceValue)
-//
-//    return AuctionDetailUi(
-//        images = photos, // photos를 사용
-//        liked = false, // isLiked 대신 기본값
-//        title = title, // manufacturer, model 대신 title 사용
-//        subTitle = "${year}년 · ${formatMileage(mileage)}",
-//        priceWon = priceValue,
-//        priceWonText = priceText,
-//        startPriceText = "",
-//        likeCount = 0, // favoriteCount 대신 기본값
-//        remainText = "",
-//        specs = createSpecsList(),
-//        notice = description ?: "",
-//        bids = emptyList(),
-//        seller = SellerUi(
-//            name = sellerInfo.name,
-//            id = sellerInfo.id,
-//            addr = "", // sellerAddress 대신 빈 값
-//            regDate = "", // sellerRegisteredAt 대신 빈 값
-//            validDate = "",
-//            count = 0, // sellerProductCount 대신 기본값
-//            response = 0, // sellerResponseRate 대신 기본값
-//            avatarUrl = null // sellerProfileImage 대신 null
-//        )
-//    )
-//}
 
-data class VehicleDetail(
-    val id: String,
-    val title: String,
-    val description: String,
-    val year: Int,
-    val mileage: Int,
-    val fuelType: String,
-    val transmission: String,
-    val engineCc: Int,
-    val horsepower: Int,
-    val color: String,
-    val price: Long?,
-    val accidentHistory: Boolean,
-    val accidentDescription: String?,
-    val photos: List<String>,
-    val options: List<String>,
-    val isSeller: Boolean,
-    val sellerId: String?,
-    val sellerName: String?,
-    val sellerLocationCity: String?,
-    val sellerProfileImageUrl: String?
-)
-
-data class SellerUi(
-    val name: String,
-    val id: String,
-    val addr: String,
-    val avatarUrl: String? = null
-)
-
-//fun VehicleDetailResponse.toVehicleDetail(): VehicleDetail {
-//    return VehicleDetail(
-//        id = id.toString(),
-//        title = "$manufacturer $model",
-//        description = description ?: "",
-//        year = year_value ?: 0,
-//        mileage = mileage ?: 0,
-//        fuelType = fuelType ?: "",
-//        transmission = transmission ?: "",
-//        engineCc = engineCc ?: 0,
-//        horsepower = horsepower ?: 0,
-//        color = color ?: "",
-//        price = price,
-//        accidentHistory = accidentHistory == "Y",
-//        accidentDescription = accidentDescription,
-//        photos = photos.map { it.photoUrl },
-//        options = options ?: emptyList(),
-//        isSeller = isSeller ?: false,
-//        sellerId = sellerId?.toString(),
-//        sellerName = sellerName,
-//        sellerLocationCity = sellerLocationCity,
-//        sellerProfileImageUrl = sellerProfileImageUrl
-//    )
-//}
 
 private fun VehicleDetail.createSpecsList(): List<Pair<String, String>> {
     val specs = mutableListOf<Pair<String, String>>()
@@ -339,16 +225,7 @@ private fun VehicleDetail.createSpecsList(): List<Pair<String, String>> {
     return specs
 }
 
-private fun formatKoreanWon(amount: Long): String {
-    val 억 = amount / 100_000_000
-    val 만 = (amount % 100_000_000) / 10_000
 
-    return buildString {
-        if (억 > 0) append("${NumberFormat.getNumberInstance(Locale.KOREA).format(억)}억 ")
-        if (만 > 0) append("${NumberFormat.getNumberInstance(Locale.KOREA).format(만)}만원")
-        if (억 == 0L && 만 == 0L) append("0원")
-    }.trim()
-}
 
 
 
@@ -379,56 +256,3 @@ fun VehicleDetail.toAuctionDetailUi(): AuctionDetailUi {
     )
 }
 
-
-//fun VehicleDetail.toAuctionDetailUi(): AuctionDetailUi {
-//    // 임시 가격 설정 (실제로는 API에서 가져오거나 계산 필요)
-//    val priceWon = 125_000_000L
-//
-//    // API 데이터로부터 스펙 생성
-//    val specs = mutableListOf<Pair<String, String>>()
-//    specs.add("연료" to fuelType)
-//    specs.add("변속기" to transmission)
-//    specs.add("배기량(cc)" to engineCc.toString())
-//    specs.add("마력" to "${horsepower}마력")
-//    specs.add("색상" to color)
-//    if (options.isNotEmpty()) {
-//        specs.add("기타 정보" to options.joinToString("\n"))
-//    }
-//    specs.add("사고 이력" to if (accidentHistory == true) "있음" else "없음")
-//    if (!accidentDescription.isNullOrBlank()) {
-//        specs.add("사고 설명" to accidentDescription)
-//    }
-//
-//    // 임의의 입찰 데이터와 판매자 정보 (실제로는 Firebase에서 가져올 예정)
-//    val bids = listOf(
-//        BidUi("홍길동", "1억 2,500만원", "2025-09-15 18:15"),
-//        BidUi("오광운", "1억 2,000만원", "2025-09-15 18:15"),
-//        BidUi("최상근", "1억 1,000만원", "2025-09-15 18:15")
-//    )
-//
-//    val seller = SellerUi(
-//        name = "태민",
-//        id = "seller123",
-//        addr = "경기 수원시 영통구",
-//        regDate = "2025.09.12",
-//        validDate = "2025.09.16",
-//        count = 39,
-//        response = 96
-//    )
-//
-//    return AuctionDetailUi(
-//        images = if (photos.isEmpty()) listOf("https://picsum.photos/id/1018/1600/900") else photos,
-//        liked = false,
-//        title = title,
-//        subTitle = "${year}년 · ${formatMileage(mileage)}",
-//        priceWon = priceWon,
-//        priceWonText = "1억 2,500만원", // 실제로는 포맷팅 필요
-//        startPriceText = "시작가 1억원",
-//        likeCount = 0,
-//        remainText = "1시간 15분",
-//        specs = specs,
-//        notice = "판매자 안내사항이 없습니다.", // API에 해당 필드가 없음
-//        bids = bids,
-//        seller = seller
-//    )
-//}
