--- app/src/main/java/com/trever/android/data/remote/dto/RefreshTokenRequest.kt
@@ -1,8 +1,0 @@
-package com.trever.android.data.remote.dto // 또는 dto 패키지 바로 아래
-
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class RefreshTokenRequest(
-    val refreshToken: String
-)
