+++ app/src/main/java/com/trever/android/ui/utils/util.kt
@@ -1,5 +1,8 @@
 package com.trever.android.ui.utils
 
+import java.text.NumberFormat
+import java.util.Locale
+
 fun formatMileage(mileageKm: Int): String {
     return if (mileageKm >= 10000) {
         val man = mileageKm / 10000
@@ -12,4 +15,15 @@ fun formatMileage(mileageKm: Int): String {
     } else {
         "${String.format("%,d", mileageKm)}km"
     }
+}
+
+fun formatKoreanWon(amount: Long): String {
+    val 억 = amount / 100_000_000
+    val 만 = (amount % 100_000_000) / 10_000
+
+    return buildString {
+        if (억 > 0) append("${NumberFormat.getNumberInstance(Locale.KOREA).format(억)}억 ")
+        if (만 > 0) append("${NumberFormat.getNumberInstance(Locale.KOREA).format(만)}만원")
+        if (억 == 0L && 만 == 0L) append("0원")
+    }.trim()
 }
